Move Transporter requests to a holding directory
