+++ src/Domain/Account.php
@@ -6,7 +6,7 @@ namespace SamuelMwangiW\Linode\Domain;
 
 use SamuelMwangiW\Linode\DTO\AccountDTO;
 use SamuelMwangiW\Linode\Factory\AccountFactory;
-use SamuelMwangiW\Linode\Request\Account\GetRequest;
+use SamuelMwangiW\Linode\Transporter\Request\Account\GetRequest;
 
 class Account
 {
