+++ src/Domain/FirewallRule.php
@@ -6,7 +6,7 @@ namespace SamuelMwangiW\Linode\Domain;
 
 use SamuelMwangiW\Linode\DTO\FirewallRulesDTO;
 use SamuelMwangiW\Linode\Factory\FirewallRulesFactory;
-use SamuelMwangiW\Linode\Request\Firewall\Rules\ListRequest;
+use SamuelMwangiW\Linode\Transporter\Request\Firewall\Rules\ListRequest;
 
 class FirewallRule
 {
