--- src/Request/Firewall/Rules/ListRequest.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SamuelMwangiW\Linode\Request\Firewall\Rules;
-
-use SamuelMwangiW\Linode\Request\LinodeRequest;
-
-class ListRequest extends LinodeRequest
-{
-}
