--- src/Request/Firewall/ShowRequest.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SamuelMwangiW\Linode\Request\Firewall;
-
-use SamuelMwangiW\Linode\Request\LinodeRequest;
-
-class ShowRequest extends LinodeRequest
-{
-}
