--- src/Request/Instance/GetRequest.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SamuelMwangiW\Linode\Request\Instance;
-
-use SamuelMwangiW\Linode\Request\LinodeRequest;
-
-class GetRequest extends LinodeRequest
-{
-}
