--- src/Request/Instance/UpdateRequest.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SamuelMwangiW\Linode\Request\Instance;
-
-use SamuelMwangiW\Linode\Request\LinodeRequest;
-
-class UpdateRequest extends LinodeRequest
-{
-    protected string $method = 'PUT';
-}
