--- src/Request/Regions/ListRequest.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SamuelMwangiW\Linode\Request\Regions;
-
-use SamuelMwangiW\Linode\Request\LinodeRequest;
-
-class ListRequest extends LinodeRequest
-{
-    protected string $path = 'regions';
-}
