+++ src/Transporter/Request/Firewall/ListRequest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace SamuelMwangiW\Linode\Request\Firewall;
+namespace SamuelMwangiW\Linode\Transporter\Request\Firewall;
 
-use SamuelMwangiW\Linode\Request\LinodeRequest;
+use SamuelMwangiW\Linode\Transporter\Request\LinodeRequest;
 
 class ListRequest extends LinodeRequest
 {
