+++ src/Transporter/Request/Firewall/Rules/ListRequest.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SamuelMwangiW\Linode\Transporter\Request\Firewall\Rules;
+
+use SamuelMwangiW\Linode\Transporter\Request\LinodeRequest;
+
+class ListRequest extends LinodeRequest
+{
+}
