+++ src/Transporter/Request/Firewall/ShowRequest.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SamuelMwangiW\Linode\Transporter\Request\Firewall;
+
+use SamuelMwangiW\Linode\Transporter\Request\LinodeRequest;
+
+class ShowRequest extends LinodeRequest
+{
+}
