+++ src/Transporter/Request/Instance/GetRequest.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SamuelMwangiW\Linode\Transporter\Request\Instance;
+
+use SamuelMwangiW\Linode\Transporter\Request\LinodeRequest;
+
+class GetRequest extends LinodeRequest
+{
+}
