+++ src/Transporter/Request/Instance/ListRequest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace SamuelMwangiW\Linode\Request\Instance;
+namespace SamuelMwangiW\Linode\Transporter\Request\Instance;
 
-use SamuelMwangiW\Linode\Request\LinodeRequest;
+use SamuelMwangiW\Linode\Transporter\Request\LinodeRequest;
 
 class ListRequest extends LinodeRequest
 {
