+++ tests/AccountTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 use SamuelMwangiW\Linode\DTO\AccountDTO;
 use SamuelMwangiW\Linode\Exceptions\CredentialsMissing;
 use SamuelMwangiW\Linode\Linode;
-use SamuelMwangiW\Linode\Request\Account\GetRequest;
+use SamuelMwangiW\Linode\Transporter\Request\Account\GetRequest;
 
 it('authenticate throws an exception when linode.config is not set', function () {
     config()->set('linode.token', '');
